Fix Windows build

+++ src/engine/serialize.cpp
@@ -461,7 +461,7 @@ std::string_view StreamBuf::toStringView( size_t sz )
     uint8_t * it2 = itget + ( sz ? sz : sizeg() );
     it2 = std::find( it1, it2, 0 );
     itget = it1 + ( sz ? sz : sizeg() );
-    return { reinterpret_cast<const char *>( it1 ), it2 - it1 };
+    return { reinterpret_cast<const char *>( it1 ), static_cast<size_t>( it2 - it1 ) };
 }
 
 void StreamBuf::skip( size_t sz )
